Todo: translate the HTML & CSS into a dynamic Next.js app. To elaborate, the README.md, index.html (ayy lmao), login.html, login_style.css, signup.html, and signup_style.css should be merged. There's nothing like moving files directly, which Next.js allows us to do to maintain the Git history continuity. See, it's been cavitated.
Todo: since Next.js allows us to copy and paste files and then npm run dev works just fine, it appears that JavaScript is an inactivated, inert mass which does not need any special compiling or hidden files... all files are visible and the JavaScript engine runs them.

Therefore, we just need to translate the HTML files into React...

https://nextjs.org/learn/basics/create-nextjs-app/setup The message is that our files can be copied and pasted, and that Next.js activation script is our guide.

Todo: It might be nice to create a globals.css, don't see why not..........

We're going to need an API URL. For the purposes of expediency, let's post this for now... localhost. When you're testing on Heroku, we will need to fetch the environment variable that is being dynamically updated whether it's localhost, herokuapp.com, or our domain pipeline which can be done via GoDaddy.

Todo: Create a file for setting environment variables that are accessible server-side...
... so that we can host our domain name therefore the application knows where to make API calls
For now, in fact, let's see if we can just make the /login and /signup page per Next.js documentation: the ./pages folder, in the format of a .js document.

Remember closing tags / !. Remember, it's style = {{}} and class goes to className when converting from HTML to JSX

As a healthy rule of thumb, kebab-case is for the HTML, camelCase is for the JSX. No one knows why...

Because Create Next.js App has provided us with a functional component, this is undefined.
Todo: Make class components so that we can access this, and then put functions on there which are bound to the context of declaration.
As for the context of invocation, that's for the arrow functions... functional components.

Bells and Whistles: Password validation; what's happening on the screen should match up with what's happening on the database, and what the server is seeing.
The onClick function is commented out for now, until we resolve it. Possible ways to resolve the onClick function are to convert the current component like /login.js and /signup.js which are in .../pages, into a Class Component which allows us to use this binding. Once we get to that stage, the onClick can be defined to access the back-end. Key take-aways: the front-end and the back-end should be hosted on the same domain. Why? Because req.body.origin is nice for access control, however it's easier to resolve the problem at its source than to resolve the symptoms of the problem. In this case, the "problem" is the fact that we want to make some kind of data fetch request, via axios library, to the back-end at which point we can unpack the request body and then make our official request to the database via the Firebase API. At this point, we can still choose our database. At any time, we can choose to use MongoDB or Firebase, depending on our wish...

Login is Signup and Signup is Login, until we can get the functionality.