Things look good so far. Having added the business plan with much-needed descriptions, the front-end has seamlessly been connected and overhauled. First, we just put the various pages in place, beforehand. Then, we might get a database set-up. We can do both, try different things out and consolidate.
Capitalization isn't a thing for macOS Terminal! That's why it's so important to allow auto-complete to do its thing, and the front-end is looking good! The class component on index.js, the imports on app.js for global CSS imports, the Next.js routing allows us to find a work-around to base url from the perspective of webpack & bundle.js, a routing system which doesn't involve putting the API_URL as an environment variable. Therefore, whatever you like to see on the website whether it's react-snowfall, react-bootstrap, these are all good things.

To add some style to it, Basic Features: Built-in CSS Support | Next.js
https://nextjs.org/docs/basic-features/built-in-css-support
Fancy CSS Borders Using Masks | CSS-Tricks - CSS-Tricks
https://css-tricks.com/css-borders-using-masks/
react-snowfall - npm
https://www.npmjs.com/package/react-snowfall. For example, in VS Code we don't turn off autocomplete; we just let it run and maybe type the title after completing the div via autocomplete. Then, we can include our business plan with ease.