Designs the signup page such that React is being imported from ‘react’, Router is being imported from ‘next/router’, and the Signup component, having been re-factored into a class component export, extends the React Component to generate constructor props based on the documentation (State and Lifecycle – React https://reactjs.org/docs/state-and-lifecycle.html) via React, the JavaScript library for building user interfaces, implementing SweetAlert (a beautiful replacement for JavaScript’s ‘alert’ (SweetAlert https://sweetalert.js.org/guides/#advanced-examples), React-Bootstrap (the library for the most popular front-end framework, rebuilt for React (React-Bootstrap · React-Bootstrap Documentation https://react-bootstrap.github.io/forms/overview/)), React-Bootstrap documentation for the Form Controls themselves (React-Bootstrap · React-Bootstrap Documentation https://react-bootstrap.github.io/forms/form-control/), and the ability to set space between the flexbox answered on GeeksforGeeks, for which we can, bar all capitalization conventions and unconventions and case sensitivity in Google Chrome, VS Code, or elsewhere, and the signupform.js created with the Router so that we can push things like /signupform, and most of all to exemplify space-between the flex box, in-line style. That is in VS Code there’s a capitalization delay for naming conventions. Case sensitivity for web browser URL can vary, so it’s so important to send the event to something first like onChange={e => console.log(‘hello”)} so that we aren’t invoking different spellings of the same filename. Conditionally disables the signup form such that the value depends on this.state, with events being sent to their target (reactjs - How do you conditionally disable a form input in react? - Stack Overflow https://stackoverflow.com/questions/48917031/how-do-you-conditionally-disable-a-form-input-in-react).
Todo, high priority:

Scrape the data from the React Bootstrap Form. Make the next steps clear; users should be able to know how to login and signup, how to use the site. Because like Chris describes, if you put yourself in the shoes of someone who is interested in chatting with a politician, the initial screen is pretty confusing. Four different options and not a single one is called out as the obvious next step… and so on and so forth. If you were interested in chatting with politicians, here is the place to be.

The idea is that we can provide the data-base as a 2nd feature so that John Doe can sign up unaffected & complemented by pre-populated data.

We’ve initialized the git repository, created the heroku app named after the repository, and initialized the nextjs-blog in order to cut and paste generated elements into the site.

Todo, low priority:

Translate the HTML & CSS into a dynamic Next.js app. Well done! With default configuration, the README.md, index.html (we need more aliens), login.html, login_style.css, signup.html, and signup_style.css have been merged, directly, since JavaScript is an inactivated, inert, and visible mass for the JavaScript engine.

We’ve created the class component. Now, let’s take functions like this.signUp and this.logIn, and the this provided by the component, and connect them together.

Login Page Functionality * Signup Page Functionality. The first thing is to create API routes on the back-end, and then make requests that reach the API routes. Then, we can set-up cookies that is one-time access tokens.

For a new politician signing up the process would be this: after they sign up (name, email, etc.) the site automatically tries to match them with a profile already created. Test user: John Doe, Democrat, Pennsylvania Senator-and that same John Doe is already in our database and has a base rudimentary profile, the first thing John will see after signing up is like a prompt that asks “is this you?”. If John says yes, he’ll have to provide more info like identity/background checks that will have to be verified. After it’s verified he’ll have full access to his already-existing profile and can add stuff like a detailed bio, make posts, etc….

The sign-up flow will be identical. “Is this you?” “Yes.” Then, the site will make a request to the database to retrieve pre-population form data. “No.” Then, the site initializes a new form… and new entry in the database…

Follow the Next.js Documentation: Now that we’ve created a global styles.css, focus on creating individualized SCSS scripts…

Want to have an index file for our database pipeline to Firebase, or MongoDB: an API URL.

Remove the business plan from the demo site.

Look into how we can effectively scrape data (for example, from ballotpedia).

Just for fun: Create a file for setting environment variables, because they are accessible server-side.

Look at Tools > Options in VS Code. Visual Studio: How to Turn Off Autocomplete - Technipages. https://www.technipages.com/visual-studio-turn-off-autocomplete