Creates the signup API which, provided that the axios HTTP request has method "POST", allows us to add new user data from the newly created new-homepage to signup-fork to signup-politician, and then with our class component is sending the data from React Bootstrap Forms (todo: our select menus should also be submitted and included), sends it to the ./pages/api folder which routes us to the database via Firebase API .collection(...).add({req.body}).
Our idea is to, like the well-known Facebook, Instagram (our color code rgb(250,250,250)), Twitter, etc. design a Vertical Layout with buttons and questions regarding whether the candidate is currently an elected official or running for office, provide a Country drop-down with one option (although now we have multiple countries with the United States at the top of the stack).

Todo: Filling out forms like the Zipcode or the Country or the Name... should pre-populate the address and pre-emptively identify the user to make it easier for them to signup. This search functionality is the same as the search functionality for signup and for the social network... we have created the form skeletons for signing up... with the message todo: Thank you for submitting your information. We will review your information for review. You will receive an email shortly to confirm your information.

Todo: Continue as Guest. * Only for elected officials, we do the full form sign-up. Pitch solution/problem 50/50 to our interested Venture Capital. Design Mangrove logo. With axios (npm install axios), we can submit to the database. Todo: design unique email verification, and grab the Excel .csv files from Google Drive (https://docs.google.com/spreadsheets/d/1SBo5A7_8Cm_nElufML8QP0KqxDIOQADb/edit#gid=479586682). Thank you, for the freeformatter.com and html-code-generator.com providing us with state and country respectively, for the select menus. For the investors, show the tree of our network for the purpose of advertising (not seed data). :)