Adds the Mathematica data-imports-and-exports.nb notebook for data editing and find & replace column values. On pages/index.js, the Login Form, the onSubmit={this.submitForm} takes the form of a submitForm function that takes the parameters async e => e.preventDefault(), and then extracts the form from e.target. Then, firebase.auth().signInWithEmailAndPassword( form.elements.email.value.toLowerCase(), form.elements.password.value.toLowerCase()) gets called before the Axios POST request to /auth/setLogin, that sets things on the current Express Session. Then, we Router.push('/userProfile'), and catch errors with an alert message. The axios.post method to /auth/setLogin is how the email from form.elements.email.value (lowercase to standardize the capitalization in Firebase) is sent from user.email to the /auth/setLogin API route, after which point the user route is pushed to /userProfile.
On the /auth/setLogin API route, var user = req.body.user. Req.session.user, which is set on the /auth/setLogin route (which has been renamed from its Qooley equivalent to the value of /sessions, because what it's doing is placing the current user email, or any value for that matter, onto the Express Session /auth/setLogin.). There, we res.status(200).json({ success: true }) in the declarative sense, and for the sake of the environment variables we are able to include them in next.config.js, which contains environment variables. However, because of the unique aspect of this site which is hosting the backend on the ./pages/api/... routes, we are able to access it directly from localhost, precluding the need for an API_URL or any kind of base URL, provided that the base URL is already localhost:3000/, or da-repo.herokuapp.com/, or anything to that extent.

* localhost:3000 says
Firebase... Enable the given sign-in provider in the Firebase console, ... Authentication > Sign-in method... Get started... Email/Password (slider). Enable
* firebaseApp.auth().signInWithEmailAndPassword (username, password) * then... There is no user record ... Signup > this.submitForm * POST /auth/setLogin /signup/saveUserToDatabase -> POST /sessions /userDatabase ✗ POST https://localhost:3000/undefined/sessions
* Changes to /api/sessions, then the response exists. ! ✗ POST ./api/userDatabase 404 (Not Found). ✓ /api/sessions is reachable, and /api/userDatabase is reachable. * ReferenceError: SweetAlert2 is not defined needs import SweetAlert2 from 'sweetalert2' * npm install sweetalert2 --save-dev * ! Hold On ! email address already in use by another account. (auth/email-already-in-use). * The Midway > Authentication > Users is where we should sign up.
* Test the Login button on localhost:3000 that is index.js. Login means nothing more than /api/sessions ... and channelling Router.push('/profilePage').
* "Login" has nothing to do with Firebase; it is the receipt of a cookie that is an access token. How are sessions sustained? Cookies.
!req.session is undefined. * npm install express-session