🍪 How to get the cookie in React | Reactgo https://reactgo.com/react-get-cookie/
Can we access the cookie from any and all components? Yes. Can we import this functional component and redesign it such that the component is just a function that returns nothing if not the cookie extracted via the "react-cookie" library? Yes; first, import the functional component: one can import ViewCookie from './ViewCookie.js' but the cookie value isn't accessible until the <ViewCookie /> component is accessible, via in-line rendering for instance.

Something that the ViewCookie component could do natively is make submissions to the API route of parameterized choice, which returns a value to the function passed as props.

function dataFetchPoliticians() {
  const data = axios.get(...)
  this.state({ data })
}

that is invoked by <ViewCookie propsFn={data...} event bubbling up in React is a thing and likewise is lifting state up; let's first hardcode the desired cookie.

Todo: On the frontend, create the pages/politician-data directory. The current /users and /loginCookies routes require access to the cookie value on mount of the ViewCookie.js component, with this binding to access new scopes like render() { return <div...}, with the useEffect(() => { propsFn(cookies) }, []) that only runs once, moved to pages/users.js, and checks whether or not the browser has a non-zero length cookie value.

🍪 Event Bubbling in React - Upmostly
https://upmostly.com/tutorials/event-bubbling-in-react

🍪 How to pass a Function as props in React | bobbyhadz
https://bobbyhadz.com/blog/react-pass-function-as-prop

🍪 Lifting State Up – React
https://reactjs.org/docs/lifting-state-up.html

🍪 react-cookie - npm
https://www.npmjs.com/package/react-cookie

🍪 reactjs - How can I set a cookie in react? - Stack Overflow
https://stackoverflow.com/questions/39826992/how-can-i-set-a-cookie-in-react

🍪 Importing a Component | Create React App
https://create-react-app.dev/docs/importing-a-component/

🍪 Using the Effect Hook – React
https://reactjs.org/docs/hooks-effect.html

🍪 Run useEffect Only Once | CSS-Tricks - CSS-Tricks
https://css-tricks.com/run-useeffect-only-once/

🍪 javascript - Correct way to handle conditional styling in React - Stack Overflow
https://stackoverflow.com/questions/35762351/correct-way-to-handle-conditional-styling-in-react