Adds GeoCodio API. That means that the curl "https://api.geocod.io/v1.7/geocode?q=1109+N+Highland+St%2C+Arlington+VA&fields=cd&api_key=166b422226b9264993154a4a4652b61423a3153" with the provided API key allows us to submit an address. Following in the footsteps of some helpful people on Stack Overflow
javascript - Best way to flatten JS object (keys and values) to a single depth array - Stack Overflow
https://stackoverflow.com/questions/44134212/best-way-to-flatten-js-object-keys-and-values-to-a-single-depth-array

javascript - One liner to flatten nested object - Stack Overflow
https://stackoverflow.com/questions/33036487/one-liner-to-flatten-nested-object

We're able to submit an address, and the tutorial (https://www.youtube.com/watch?v=ogFpMIw6ugo) is there a way to see responses on Discord? Yes, we can Jump to Original Message. We want to be able to submit an address and now we can, it'll still return something as long as we have a city or state. For delegates we get contact information... a traditional example os the three representatives from where we're living: Congress Search. Going through the tutorial, for which we generate a Rails app, installing with sudo permissions in the /usr/local/bin with the option -n, we are able to generate the Rails application as an API, updating the Gem file and installing NokoGiri, and updating the Ruby version within the code itself. Then, we secure the API key, get some resources, and then create some controller action. We have some Routing, and this is so that we can hook up our back-end to the GeoCodio API. This commit is just a test.

The video tutorial uses Ruby, because it roughly does something similar to what we want we are going to replace the zip code of the user with address, address city state etc. because apparentlyl zip codes are unreliable to get district info. Large cities like Pittsburgh/Philadelphia have things called "wards" led by an alderman. So we're able to install the sudo gem install nokogiri -n /usr/local/bin with HomeBrew and get all the district info, passing address data to their website and getting their results back, in the console.