What could be causing the infinite loop? It’s possible that this.setState({ ‘responseSuccess’: true }) that is what causes the infinite redirect loop, effectively however there’s also the rendering of <LoginCookies and <ProfilePage within the component pages/index.js.
The infinite loop is initiated when the setState responseSuccess coincides with the re-mounting of the component. There is an <input onChange={e => this.setState({ username: e.target.value })}, which coincides with the development of a infinite loop because the onChange method keeps getting triggered indefinitely once we click Login.

What we could do is set the login function to disable the login functionality for say 10 seconds until we are able to log in properly that is, without entering an infinite loop.

This “necessitates” the creation of a helper function. For this we can call the functions known as this.setUsername(e.target.value, and this.setPassword(e.target.value. Therefore we have set functions bound to this.state. Is this enough to prevent the dreaded infinite feedback loop?  Do events come with a pre-defined e.preventDefault() behavior?

Is setting the if statement condition to true enough, to set things on state without achieving an infinite feedback loop? The answer is an emphatic no!

The feedback loop may have multiple causes; that setting state immediately in-line instead of setting it within a helper function, and now, the second cause: <LoginCookies and/or <ProfilePage components. Let’s comment out the LoginCookies component. This doesn’t result in any infinite loop! The ProfilePage component is not the cause of the infinite loop, the LoginCookies component is the cause. This conjecture is corroborated by the evidence that the LoginCookiesOld component is console logging that we are on the login cookies component*****, via the useEffect hook. Why does the LoginCookiesOld component keep getting mounted? The error is possibly in the fact that we are having two useEffect hooks. Is this customary? The answer is no, because the first useEffect loop, when commented out does nothing to assuage the useEffect hook.

What part of the useEffect Hook is causing the infinite loop? Yes, it is the second useEffect asynchronous function that causes the infinite loop. Perhaps it is the Router.push command inherent in the LoginCookiesOld component.

Absolutely not. Commenting out Router.push(‘/‘) makes the infinite loop even faster than before! Is the loop caused by the act of setting a cookie? Yes. This should be obvious; the fact is that we are setting a cookie using React hooks, and using the “react-cookie” library which inherently makes use of the existing React Hooks, we need to determine if we’ve already set a cookie before. What we want to do is set a cookie once, but also set a variable preventing the infinite loop from occurring. Furthermore, this variable setting takes place at once therefore won’t interfere more than once with the React Lifecycle. Why does the variable “do” and the corresponding “setDo”. We’ve also got the const [variable, setVariable] = React.useState(0); this works quite nicely. Ideally, if we setVariable(1) in the if (cookieSet && variable == 0), we are able to set a one-time trigger for setting the cookie on the user’s browser. The only thing left to do is trigger the browser refresh, which may be natively built into the browser. This is very successful, and now we would like to refresh the browser a little later so that it has time to fetch the cookie from the server.

The key is to call the vanilla JavaScript *reload* method to tell the browser to reload the current page. window.location.reload(true);. It requires at least 3000 milliseconds for the page to refresh, and simultaneously we can set the cookie with a maximum age of 20 seconds. Implementing with this delay, we have produced a simple login system. If you want to access the data on our site, go through our login system with the username: haha@haha.com password: hahahaha. Therefore we have moved the address search to the ProfilePageOld.js component.

❤️How to Refresh a Page or Component in React - Upmostly
https://upmostly.com/tutorials/how-to-refresh-a-page-or-component-in-react

❤️Using the State Hook – React
https://reactjs.org/docs/hooks-state.html

❤️javascript - Preventing onchange events from getting into an infinite loop - Stack Overflow
https://stackoverflow.com/questions/8833778/preventing-onchange-events-from-getting-into-an-infinite-loop

❤️TIL: How to pause an infinite JavaScript loop in Chrome - DEV Community 👩‍💻👨‍💻
https://dev.to/ajones_codes/til-how-to-pause-an-infinite-javascript-loop-in-chrome-53gm

❤️How to open the Task Manager on Mac: monitoring apps and processes - IONOS
https://www.ionos.com/digitalguide/server/configuration/task-manager-mac/#:~:text=The%20Mac%20Task%20Manager%20is,are%20running%20in%20the%20background.