Unfortunately the react-cookie package does not allow you to simply delete all cookies. What you can do, is find the name of the cookies, and that’s where things get complicated.
Is there a way to see all the cookies that are available on the Chrome Web Browser? Yes. On the above documentation, we are able to access the ‘cookies’ object. What do we do with this object? We console.log it. Since we are able to set a timeOut of 3000 milliseconds that is 3 seconds on the LoginCookiesOld component as the threshold of time elapsed before refreshing the page, then we are also able to set a cookie with a maximum age of 60 seconds, that is the cookie signifying login can be set with a maximum age, of 60 seconds. This is in LoginCookiesOld.js, whereas in ProfilePageOld.js where the component updates we are able to derive the this.state.cookieValue.user.data cookie. If we wanted to log out, we would just need two things. Refresh the page, verify the cookie persistence across browser refreshes. This is verifiable, see for yourself! What we need to do first is delete the cookie off the browser and then initiate a refresh. Those two actions will constitute logging out. The name of this cookie is simply, “user”. This is evident in the Applications tab of the Chrome DevTools console; instead of setCookie, we want to invoke the removeCookie function. The interesting thing is that the cookie is being fetched from a lower level hierarchy component, a function that sets the traditional state that is ProfilePageOld.js, with the context of invocation being the LoginCookiesOld component.

Therefore the Login Button could be made as a lower level functional component. We are successfully able to import the LogoutComponent with its button and the fact that we see this button means that we are on the right track. Therefore, are we using the useEffect hook on the LogOut component? Yes, we are. By re-initiating the library react-cookie npm import, we are able to understand the origin of the cookie; it’s already on the browser, rather than being defined by the component it’s “instantiated” in. Then, what we can do is take the value that we have retrieved and then invoke the removeCookie function. Can we do this outside of the react-cookie API, with the browser API? Yes? Set the cookie to be expired after 600 seconds while we are testing the cookie functionality that is the logout browser.cookies.remove(“user”). We can name this cookie to anything we want.

Everything is centered at the cookie name in the code, while the cookie is still specified as a proper parameter in the react-cookie library. That is, everything is based on the cookie name.

javascript - How to remove cookies completely using useCookies in React without leaving undefined - Stack Overflow
https://stackoverflow.com/questions/68600015/how-to-remove-cookies-completely-using-usecookies-in-react-without-leaving-undef

react-cookie - npm
https://www.npmjs.com/package/react-cookie

cookies - Mozilla | MDN
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/cookies