Things are looking great. The Logout function is neat and the button style is wild. Now we need a signout function; renames the button to “Logout”. Want to place it in the location of what would be the Login page, if we weren’t logged in.
Where are the styles of the Login Button? Where is the Login page being used? It definitely is being used in the .next/ directory… in it there’s the lottie-player HTML tag that has src=“https://assets4.lottiefiles.com...profile.json” value although the component when accessed in the web browser redirects to the main page ./pages/index.js.

The Mangrove at https://themangrove.vercel.app/ is something to live for; it's the kind of place you want to live in because it is like a labyrinth, a puzzle, a place to be explored because it draws in electricity all by itself. Thank you for suggesting that we can make a signout function. It means a lot.

The rendering on ProfilePageOld of the LogoutComponent which is the rendering where the conditional server response data is required to render anything aside from the initial Login form. To put it a different way, if the Firebase server responds then we don’t need a login form. The Login form is in another component, therefore if we get any kind of data on the ProfilePageOld component which is a simple ternary then we can { this.state.data && <LogoutComponent /> } and then get our LogoutComponent. The interesting thing is that the data is there whether or not the user is logged in, that is the server always responds. Therefore the data is for the time being available on this.state.data because Firebase responds. Therefore we can just { this.state.cookieValue && <LogoutComponent /> } and then furthermore, because this.state.cookieValue is present whether or not we are logged in, we must also check for this.state.cookieValue && Object.keys(this.state.cookieValue).length !== 0. By contrast Object.keys(this.state.cookieValue).length === 0 is a binary query which scatters the essence of knowledge that the user is logged out, and what about rendering the Login Page only when the user isn’t logged in? What’s good is that the LoginCookiesOld.js component isn’t rendering the Login Form; otherwise, we would realize that the pages/index.js page is rendering the Login Form in a straightforward fashion.

On pages/index.js, would we know if we were logged in or not? Quite possibly, we would know by looking at the this.state.responseSuccess variable that refers to a value that is set via the axios.post /api/sessions route. But can we access this.state.responseSuccess elsewhere? Yes, and as a matter of fact we should. Also, adds a lot of color and expansive style to the Logout button, if you can’t even see it then there’s no point.

via | Etymology, origin and meaning of via by etymonline
https://www.etymonline.com/word/via

ablative | Etymology, origin and meaning of ablative by etymonline
https://www.etymonline.com/word/ablative

Bubble coloring button
https://codepen.io/comehope/pen/eKqZjy

35 CSS Button Hover Effects
https://freefrontend.com/css-button-hover-effects/

28 React JS Buttons
https://freefrontend.com/react-js-buttons/

coolest button styles in react - Google Search
https://www.google.com/search?q=coolest+button+styles+in+react&oq=coolest+button+styles+in+react&aqs=chrome..69i57j33i160l2j33i299.2745j0j7&sourceid=chrome&ie=UTF-8