Todo: tidy up the data, maybe make like a few buttons on the home screen that'll link to other pages, after you log in.
To look at the contrary, no need to pigeon-hole this product. Don't need to allow voters to communicate anything, they just need to communicate.🙂 Our project is about spoofing, subscribing, and the love of life. With our team, we are walking in the footsteps of giants.

Non-politicians should be anonymous by default. Todo: add general assembly members to the search function.

This is our time. npm start. With Node Package Manager installed, npm install, run npm start to start a Node server. Within a few seconds, your browser should load the frontend. You're ready to start searching! They want us to participate and make our voice heard. In this tutorial, we run both a Rails server and npm start so that the backend and frontend can communicate.

When you try to search with just the front-end, you'll get an ERR_CONNECTION_REFUSED on port :3001. In order to rails db:migrate you need to install nokogiri. First, we need Nokogiri.

When we have successfully installed sudo gem install nokogiri -n /usr/bin, then we will be done installing the documentation for nokogiri after 1 seconds. Then, we can bundle update nokogiri with the ignoring commonmarker-0.17.13 because its extensions are not built... using concurrent-ruby 1.1.10... the bundle is updated!

We found that Zip Codes and District Info like Wards are hard to get information on, they're made along different lines of thought! We can replace the Zip Code of the User with Address, City, State, etc. 🚋

Remember to git pull fork-origin master! You may not know when it's updated.

Comments out background radial gradient with the circles and transparent properties, to show an orange themed background color.

We could use the value of this.state.responseSuccess && as a criterion for displaying this.state.flattenedAddress... however, this.state.responseSuccess exists when the user is currently logging in, which may not always be the case.

We can also use this.state.cookieValue && as a criterion for displaying this.state.flattenedAddress and mapping keys to values. It's a "cookieValue".😝

It's from dataFetchPoliticians, intended to be a part of the politicians' data that is set directly on the cookie. From this, we can execute the propsFn which is this.dataFetchPoliticians, within the CookieView component. That is we are fetching data for the politicians and passing this function, which puts things on this.state in the parent component, as a property function to the CookieView component that views cookies. Is this a cookie value? No! It's present regardless of status logged-in.

😄 we can cd congress, press Tab, and auto-complete will give us cd code-sample-react-rails-congress-search! origin    https://github.com/Geocodio/code-sample-react-rails-congress-search.git (fetch)

Sources:

Installing Nokogiri - Nokogiri
https://nokogiri.org/tutorials/installing_nokogiri.html#supported-platforms

How do I fix the "You don't have write permissions into the /usr/bin directory" error when installing Rails? - Stack Overflow
https://stackoverflow.com/questions/2893889/how-do-i-fix-the-you-dont-have-write-permissions-into-the-usr-bin-directory

ruby - 'require': cannot load such file -- 'nokogiri\nokogiri' (LoadError) when running rails server - Stack Overflow
https://stackoverflow.com/questions/28999906/require-cannot-load-such-file-nokogiri-nokogiri-loaderror-when-running