* Added some sort of google maps functionality to the show users part. That's a decent step, are you gonna be able to extract lat/long from your location that way too? chrome://settings/content/location * `http://localhost:3000 wants to 📍 Know your location (Block) (Allow)` Yoinks you to the Pfingsten area (without a marker). * The in MapContainer.js * useEffect(() => { navigator.geolocation.getCurrentPosition(success);... * ... `navigator` happens once we `allow` location in Chrome. The `useEffect` function keeps looping once: redirects our view to the [ currentPosition, setCurrentPosition ] React Hook. The `success` "variable" that only gets set once.
It looks like adding console logs makes the map think twice before (doesn't) jump to static, the current position. Back to the original question we can extract lat/long ✓ and can we put it on a marker? ? How does GoogleMap address the child divs <Marker...? Yes, we are going to "be able to extract lat/long" that's beyond question.
Task 2. Need to "somehow" link that to "ArcGIS" using `the method "I" mentioned earlier` - that is the most probable method
of the ancients.
Buried in the lost sands of time, ArcGIS API for JavaScript: `developers.arcgis.com/documentation/mapping-apis-and-services/search/reverse-geocoding/` npm install @arcgis/core
. The keys are being mapped `to` `from` { this.state.`flattenedAddress`[key]}
Like likes like. "the response has been received!".
`this.state.latLngQueryFlattened` is just like `this.state.flattenedAddress;` The search query parameter object requirements are out of this world! "What" are they, are they on the ArcGIS documentation? That is for latitude & longitude, and address.