Where is the request http://localhost:5000/posts being made It's only being made once The idea is that "something" is supposed to be hosted on `const url = 'http://localhost:5000/posts';` No database server! That means there must be a /server running. What if the `memories_project` could just "Run Alongside" the existing `Da-Repo`? What is the object that gets sent? Why is "Delete Immediately" shown in the Trash, on macOS? Instead of the Creator Title Message Tags Choose File ordering.. Present on the initial "memories project", we have a lot of messages "being posted" but we have them on Creator Message SelectedFile tags and title. Can we match these to what happens on the screen? `client/src/api/index.js` -- memories_project ```import axios from 'axios';
const url = 'http://localhost:5000/posts';

export const fetchPosts = () => {
  console.log("We are fetching with the 'input'", url)
  axios.get(url)
};
export const createPost = (newPost) => {
  console.log("The url is what is it? provided in the global context. The real input we have is ", newPost)
  axios.post(url, newPost);
}
export const updatePost = (id, updatedPost) => {
  console.log("We can still update the post but we do it with id", id, "and updatedPost ", updatedPost)
  axios.patch(`${url}/${id}`, updatedPost);
}
export const deletePost = (id) => {
  console.log("We are about to delete a post at this Id", id)
  axios.delete(`${url}/${id}`);
}```
`http://localhost:3000/users?name=null`
In `helpers/actions/postsAPI.js` -- Da-Repo There is something glorious.
"Now" try to upload a file
When we try to upload a file, `setPostData` is not defined. . .
The request being made to the back-end ... it's just an object!::!:
```The url is what is it? provided in the global context. The real input we have is  null```
```Cannot destructure property 'data' of '(intermediate value)' as it is undefined.```
In `postsAPI.js` there is the connection to the back-end being made via `axios`.
In `posts.js` there is the connection to the back-end, being made via `Cannot destructure property 'data' of '(intermediate value)' as it is undefined.` via CreatePost in `helpers/actions/posts.js`
The data object in `Cannot destructure property 'data' of '(intermediate value)' as it is undefined.` of `const { data } = await api.createPost(post);` in `helpers/actions/posts.js` is `undefined`!