Expands upon the list of Python commands and packages needed for web scraping and data analysis. That includes installing Spacy and NLTK, two popular natural language processing libraries, as well as Pandas and Seaborn for data manipulation and visualization. We start out with the regular packages for web scraping, such as Beautiful Soup and Requests, and the instructions to download the en_core_web_sm model for Spacy. What's new? The command pip3 install openpyxl installs the openpyxl package which is a Python library for reading and writing Excel 2010 xlsx/xlsm/xltx/xltm files. This allows us to write data to Excel sheets, as well as read data from them. This package can be useful for automating tasks that involve working with Excel files.
We're web scraping county links from a Wikipedia page, using requests and BeautifulSoup libraries. We then use openpyxl to create a new Excel workbook and write the extracted county names and links to the workbook's first worksheet. The final workbook is saved as an XLSX file.

Here we've got the Python script that scrapes data from the website https://ilga.gov/house/ to extract the names and district numbers of all members of the Illinois House of Representatives.

The script uses the requests and BeautifulSoup libraries to scrape the website, find the relevant HTML elements, and extract the data. It then uses the openpyxl library to create a new Excel workbook and worksheet, add column headers, and write the extracted data to the worksheet.

The final output of the script is an Excel file named "ilga_members.xlsx" that contains the names and district numbers of all members of the Illinois House of Representatives.

Scrapes the county information from https://www.ilsheriff.org/sheriffs-directory/ using BeautifulSoup and requests libraries in Python. It extracts the county name, phone number, name, location, and email information for each county. It also writes the extracted information to a CSV file named "sheriffs.csv" using the csv library in Python. The extracted information is written to the CSV file in columns - "County Name", "Phone Number", "Location", and "Email".

Imports requests and pandas libraries and uses BeautifulSoup for web scraping. Adds code to scrape the Illinois Sheriff's directory to extract data for all counties. Defines a function scrape_county_information() to extract county name, phone number, location, and email from each county's page. The main code then iterates through the first three counties, extracts the relevant information using the scrape_county_information() function, and appends it to a list of data. Finally, converts the list to a Pandas DataFrame and saves the data as an Excel file named 'sheriffs.xlsx'.